Update ep_gui.cs-CZ.rc
addead few translations from the bottom as a test. cannot find ability to comment. there is no tool to settle on terms like taskbar etc. I prefer to continue on Google Docs. It's one more step, but we are not programming here, we just edit text, so it could be better